main code
This is the original code that implements the approach published on Transactions on Industrial Informatics titled "Knowledge Distillation for Scalable Non-Intrusive Load Monitoring" by Tanoni et al. ( doi: 10.1109/TII.2023.3328436).

The network structure is a CRNN pre-trained both with UKDALE and REFIT dataset, and then fine-tuned to be tested on REFIT houses. Both pretrained and finetuned models are available.